requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
bidict==0.23.1
cachetools==6.2.6
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
cryptography==46.0.7
deprecation==2.1.0
fastapi==0.136.0
fsspec==2026.3.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.12
markdown-it-py==4.0.0
mdurl==0.1.2
mmh3==5.2.1
multidict==6.7.1
packaging==26.1
postgrest==2.28.3
propcache==0.4.1
pycparser==3.0
pydantic==2.13.3
pydantic_core==2.46.3
Pygments==2.20.0
pyiceberg==0.11.1
PyJWT==2.12.1
pyparsing==3.3.2
pyroaring==1.0.4
python-dateutil==2.9.0.post0
python-engineio==4.13.1
python-socketio==5.16.1
realtime==2.28.3
requests==2.33.1
resend==2.29.0
rich==14.3.4
simple-websocket==1.1.0
six==1.17.0
starlette==1.0.0
storage3==2.28.3
StrEnum==0.4.15
strictyaml==1.7.3
supabase==2.28.3
supabase-auth==2.28.3
supabase-functions==2.28.3
tenacity==9.1.4
typing-inspection==0.4.2
typing_extensions==4.15.0
upstash-redis==1.7.0
urllib3==2.6.3
uvicorn==0.45.0
websockets==15.0.1
wsproto==1.3.2
yarl==1.23.0
zstandard==0.25.0
